+++ lib/src/platform.dart
@@ -11,17 +11,17 @@ class Platform {
   /// The method returns a [Future] that completes with a [String] containing
   /// the platform of the device, or a [bool] indicating if the device could
   /// not be queried.
-  static Future<dynamic> get(ZKTeco self) async {
+  static Future<String?> get(ZKTeco self) async {
     int command = Util.CMD_DEVICE;
     String commandString = '~Platform';
 
-    dynamic reply = await self.command(command, commandString);
+    var resp = await self.command(command, commandString);
 
-    if (reply is bool) {
-      return reply;
+    if (resp['status'] == false) {
+      return null;
     }
 
-    return String.fromCharCodes(reply);
+    return String.fromCharCodes(self.dataRecv.sublist(8, 28));
   }
 
   /// Returns the version of the device as a [String].
@@ -37,12 +37,8 @@ class Platform {
     int command = Util.CMD_DEVICE;
     String commandString = '~ZKFPVersion';
 
-    dynamic reply = await self.command(command, commandString);
+    await self.command(command, commandString);
 
-    if (reply is bool) {
-      return reply;
-    }
-
-    return String.fromCharCodes(reply);
+    return String.fromCharCodes(self.dataRecv.sublist(8, 23));
   }
 }