+++ test/flutter_ztkteco_test.dart
@@ -6,7 +6,6 @@ import 'package:flutter_zkteco/flutter_zkteco.dart';
 void main() async {
   test('connect & disconnect from machine', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     bool? isConnected = await fingerMachine.connect();
 
     expect(isConnected, true);
@@ -19,7 +18,6 @@ void main() async {
   test('enable live attendance', () async {
     final fingerMachine =
         ZKTeco("10.7.0.53", timeout: const Duration(minutes: 1));
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
 
     await fingerMachine.enableLiveCapture();
@@ -34,7 +32,6 @@ void main() async {
 
   test('get version success', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic version = await fingerMachine.version();
     expect(version is String, true);
@@ -43,7 +40,6 @@ void main() async {
 
   test('get version failed', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic version = await fingerMachine.version();
     expect(version is String, false);
@@ -52,7 +48,6 @@ void main() async {
 
   test('get serial number', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic serialNumber = await fingerMachine.serialNumber();
     expect(serialNumber is String, true);
@@ -61,7 +56,6 @@ void main() async {
 
   test('get device name', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic deviceName = await fingerMachine.getDeviceName();
     expect(deviceName is String, true);
@@ -70,7 +64,6 @@ void main() async {
 
   test('get platform', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic platform = await fingerMachine.platform();
     expect(platform is String, true);
@@ -79,7 +72,6 @@ void main() async {
 
   test('get platform version', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic platformVersion = await fingerMachine.platformVersion();
     expect(platformVersion is String, true);
@@ -88,7 +80,6 @@ void main() async {
 
   test('get os', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic os = await fingerMachine.getOS();
     expect(os is String, true);
@@ -97,7 +88,6 @@ void main() async {
 
   test('get datetime', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic time = await fingerMachine.getTime();
     expect(time is String, true);
@@ -106,7 +96,6 @@ void main() async {
 
   test('set datetime', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic time = await fingerMachine.setTime(DateTime.now());
     expect(time is String, true);
@@ -115,7 +104,6 @@ void main() async {
 
   test('get SSR', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic ssr = await fingerMachine.getSsr();
     expect(ssr is String, true);
@@ -124,7 +112,6 @@ void main() async {
 
   test('get Users', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     List<UserInfo> users = await fingerMachine.getUsers();
     // ignore: unnecessary_type_check
@@ -134,7 +121,6 @@ void main() async {
 
   test('get Attendances', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     List<AttendanceLog> logs = await fingerMachine.getAttendanceLogs();
     // ignore: unnecessary_type_check
@@ -144,7 +130,6 @@ void main() async {
 
   test('test voice', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
     await fingerMachine.connect();
     dynamic platform = await fingerMachine.testVoice();
     expect(platform is String, true);
@@ -153,7 +138,6 @@ void main() async {
 
   test('test restart', () async {
     final fingerMachine = ZKTeco("10.7.0.53");
-    await fingerMachine.initSocket();
 
     await fingerMachine.connect();